pyproject.toml
[project]
name = "archon-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.2.6",
    "duckduckgo-search>=2025.4.1",
    "exa-py>=1.10.0",
    "fastapi>=0.115.12",
    "google-genai>=1.9.0",
    "jinja2>=3.1.6",
    "mcp>=1.6.0",
    "nltk>=3.9.1",
    "pydantic>=2.11.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "spacy>=3.8.5",
    "uvicorn>=0.34.0",
]
